requirements.txt
streamlit==1.33.0
pillow==10.2.0
